% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prediction.R
\name{add_aggr_source_type}
\alias{add_aggr_source_type}
\title{add an aggregation function to a pollyvote object}
\usage{
add_aggr_source_type(pv, method, which_source_type, agg_fun = "mean",
  na.handle = "na.rm", ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the data to.}

\item{method}{[\code{character(1)}]\cr
method of the prediction. This method name will be used to call the prediction in \code{\link{predict.pollyvote}}.}

\item{which_source_type}{[\code{character(n)}]\cr
character vector of arbitrary length containing the \code{source_type}s 
over which to aggregate.}

\item{agg_fun}{[\code{character(1)}]\cr
string indicating which aggregation function to use. Currently implemented 
are 'mean' and 'median'.}

\item{na.handle}{[\code{character(1)}]\cr
string indicating which aggregation function to use. Currently implemented 
are 'na.rm', indicating that missing observations will be ignored and all
other functions will be aggregated and TODO.}

\item{...}{additional arguments}
}
\value{
The pollyvote object with added prediction
}
\description{
Adds an aggregation function to a pollyvote object. An aggregation function is
a special type of a prediction function.
}
