% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data.pollyvote}
\alias{add_data.pollyvote}
\title{add data to a pollyvote object}
\usage{
\method{add_data}{pollyvote}(pv, newdata = data.frame(), ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the data to.}

\item{newdata}{[\code{data.frame}]\cr
the data to add to the pollyvote object.}

\item{...}{additional arguments giving information about \code{newdata}.}
}
\value{
The pollyvote object with added data.
}
\description{
Adds data to a pollyvote opbject.
}
