% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{add data (to a pollyvote object)}
\usage{
add_data(pv, newdata = data.frame(), ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the data to.}

\item{newdata}{[\code{data.frame}]\cr
the data to add to the pollyvote object.
should contain this columns: id, date, source, party, percent}

\item{...}{additional arguments giving information about \code{newdata}.}
}
\value{
The pollyvote object with added data.
}
\description{
Adds data, so far only implemented for pollyvote objects.
}
\examples{
data("polls_individual")
pv = add_data(pv, newdata = polls_individual, country = "D", 
              region = "national", source_type = "poll", election = "BTW")

}
