% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_election_result.R
\name{add_election_result.pollyvote}
\alias{add_election_result.pollyvote}
\title{add an election result to a pollyvote object}
\usage{
\method{add_election_result}{pollyvote}(pv, election, data, ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the election results to.}

\item{election}{[\code{character(1)}]\cr
name of the election. One of the permitted elections of \code{pv}.}

\item{data}{[\code{data.frame()}]\cr
data frame containing the election results in the long format.}

\item{...}{currently ignored.}
}
\value{
The pollyvote object with added prediction
}
\description{
Adds election result to a pollyvote opbject.
The election date must be in format "%Y-%m-%d" and of type POSIXlt', 'POSIXct' or 'character'.
}
\examples{
pv = create_pollyvote(perm_countries = "D") 
data("election_result")
pv = add_election_result(pv, "BTW 2013", election_result)

}
