% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_election_result.R
\name{add_election_result}
\alias{add_election_result}
\title{add election results (to a pollyvote object)}
\usage{
add_election_result(pv, election, data, ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the election results to.}

\item{election}{[\code{character(1)}]\cr
name of the election. One of the permitted elections of \code{pv}.}

\item{data}{[\code{data.frame()}]\cr
data frame containing the election results in the long format.}

\item{...}{currently ignored.}
}
\value{
The pollyvote object with added prediction.
}
\description{
Adds an election result, so far only implemented for pollyvote objects.
}

