% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_error_calc.R
\name{add_error_calc.pollyvote}
\alias{add_error_calc.pollyvote}
\title{add data to a pollyvote object}
\usage{
\method{add_error_calc}{pollyvote}(pv, method = "TODO", fun = function(pv) {
      stop("TODO") }, ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the data to.}

\item{method}{[\code{character(1)}]\cr
name of the error calculation.}

\item{fun}{[\code{function(pv)}]\cr}

\item{...}{additional arguments}
}
\value{
The pollyvote object with added prediction
}
\description{
Adds data to a pollyvote opbject.
}
\examples{
pv = create_pollyvote(perm_countries = "D")
pv = add_error_calc(pv, "poll_only", function(pv) {
  pred_data = predict(pv, "poll")
  result = get_election_result(pv, "BTW")
  joined = left_join(x = pred_data, y = result, by = "party") \%>\%
    rename(percent = percent.x, percent.true = percent.y)
  return(mutate(joined, error = abs(percent - percent.true)))
})

}
