% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_error_calc.R
\name{add_error_calc}
\alias{add_error_calc}
\title{add an error calculation (to a pollyvote object)}
\usage{
add_error_calc(pv, method, fun, ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the data to.}

\item{method}{[\code{character(1)}]\cr
name of the error calculation.}

\item{fun}{[\code{function(pv)}]\cr}

\item{...}{additional arguments}
}
\value{
The pollyvote object with added prediction.
}
\description{
Adds a prediction, so far only implemented for pollyvote objects.
}
