% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prediction.R
\name{add_prediction.pollyvote}
\alias{add_prediction.pollyvote}
\title{add data to a pollyvote object}
\usage{
\method{add_prediction}{pollyvote}(pv, method = "no_aggregation",
  fun = function(pv) {     get_data(pv) }, ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the data to.}

\item{method}{[\code{character(1)}]\cr
method of the prediction. This method name will be used to call the prediction in \code{\link{predict.pollyvote}}.}

\item{fun}{[\code{function(pv)}]\cr}

\item{...}{additional arguments}
}
\value{
The pollyvote object with added prediction
}
\description{
Adds data to a pollyvote opbject.
}
\examples{
pv = create_pollyvote(perm_countries = "D")
pv = add_prediction(pv, "poll", function(pv) {
  pv \%>\% 
    get_data \%>\% 
    filter(source_type \%in\% c("poll")) \%>\%
    group_by(date, source_type, party) \%>\% 
    summarize(percent = mean(percent, na.rm = TRUE))
})

}
