% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_region_weights.R
\name{add_region_weights}
\alias{add_region_weights}
\title{Adds the region weights to the pollyvote container.}
\usage{
add_region_weights(pv, region_weights)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add region weights to.}

\item{region_weights}{[\code{data.frame}]\cr
Data frame with columns 'region' and 'weight'
representing region names and corresponding region weights.
Region names are validated against the permitted regions in the pollyvote container.

@examples 
pv = add_region_weights(pv, region_weights)

@return the pollyvote object with added region weights

@export}
}
\description{
Adds the region weights to the pollyvote container.
}
