% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_regions.R
\name{add_regions}
\alias{add_regions}
\title{Adds the region weights to the pollyvote container.}
\usage{
add_regions(pv, regions, weights = NULL)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add region weights to.}

\item{regions}{[\code{character(n)}]\cr
Vector of region names
Region names are validated against the permitted regions in the pollyvote container.}

\item{weights}{[\code{numeric(n)}]\cr
Vector of region weights. If \code{NULL}, then all weights are assigned to 1 (Default value).}
}
\value{
the pollyvote object with added region weights
}
\description{
Adds the region weights to the pollyvote container.
}
\examples{
  pv = add_regions(pv, c("London", "Yorkshire"), c(0.7, 0.3))
  pv = add_regions(pv, c("London", "Yorkshire"))
  
}
