% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_by.R
\name{aggreg_by}
\alias{aggreg_by}
\title{aggregate percentage values}
\usage{
aggreg_by(data, group_variab, aggr_func = mean)
}
\arguments{
\item{data}{[\code{data.frame(n)}]\cr
data.frame to be aggregated}

\item{group_variab}{[\code{character(n)}]\cr
grouping variables}

\item{aggr_func}{[\code{character(n)}]\cr
aggregation function}
}
\value{
aggregated data.frame
}
\description{
This is a function that aggregate percent values by groupping
variables group_variab.
}
\examples{
elect_data_wide <- import_exp_data_wide("Wahlumfrage")
elect_data_long <- wide_to_long_format(elect_data_wide)

elect_data_aggr_mean <- aggreg_by(elect_data_long, group_variab = c("date", "party"), aggr_func = mean)
elect_data_aggr_max <- aggreg_by(elect_data_long, group_variab = c("date", "party"), aggr_func = max)


}
