% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalitions.R
\name{calc_coalitions}
\alias{calc_coalitions}
\title{Coalitions percentage prediction.
 
Provides prediction of percentages from different source types for specified coalitions. 
Coalitions percentages are predicted for each day in range [election_day - limit_days, election_day).}
\usage{
calc_coalitions(pv, coalitions, threshold = 0, threshold_handle = "omit",
  prediction = "pollyvote", ..., election_year = NULL,
  permitted_parties = NULL, limitdays = -1, for.ggplot2 = FALSE)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr 
The pollyvote object of which to get the prediction from.}

\item{coalitions}{[\code{list}]\cr
List of vectors representing coallitions. Coallitions must be specified with full and exact party names.}

\item{threshold}{[\code{numeric(1)}]\cr
If positive, indicates the minimum voice component
that a party has to reach in order to participate in a coalition.}

\item{threshold_handle}{[\code{character(1)}]\cr
Specifies how to handle coalitions with parties that have less percentage than threshold.
Options:
\itemize{
 \item \code{'omit'}: default value. In this case, this coalition percentage is not calculated for the given date. Instead, NA is entered.
 \item \code{'ignore'}: In this case, the coalition will be made up of the remaining parties. 
 }}

\item{prediction}{[\code{character(1)}]\cr
Name of the prediction function.
For calculation of coalitions aggregated over (date, source_type, party) [\code{pollyvote}] should be supplied.
For calculation of coalitions aggregated over (region, date, source_type, party) [\code{aggr_pollyvote_region}] should be supplied.
For calculation of coalitions only in one region [\code{pollyvote_region}] should be supplied.
Options:
\itemize{
 \item [\code{character(1)}]: Single component name (e.g., "poll"); in this case will one column for the specified component
   and one column each (direct) Subcomponent returned.
 \item [\code{character(n)}]: Vector with component names - gives exactly for the specified components (without automatic subcomponent addition) the values back.
 \item [\code{NULL}]: Wildcard that returns all components.
 }}

\item{...}{Optional parameters passed into [\code{prediction}] function.
This function needs to be called with region = 'region' in the optional params if the results need to be calculated for one region only.}

\item{election_year}{[\code{numeric(1)}]\cr
The election year for which the coalitions are predicted.
If not specified, the most recent election year is used.}

\item{permitted_parties}{[\code{character(n)}]\cr
Selection of only specific parties for which coalitions are calculated.
Options:
\itemize{
 \item [\code{NULL}]: Default value for not excluding any party in coalitions calculations.
 \item [\code{character(1)}]: For specifying one party.
 \item [\code{character(n)}]: Vector for specifying multiple parties.
 }}

\item{for.ggplot2}{[\code{logical(1)}]\cr
Return format of coalitions predictions.
Options:
\itemize{
 \item [\code{logical(1)}] FALSE: Default value - Function returns data frame of columns (date | Days to election | Coalition_1_percentage | ... | Coaltion_n_percentage)
 \item [\code{logical(1)}] TRUE:  Data frame with rows containing visualisation points
 }}

\item{limit_days}{[\code{numeric(1)}]\cr
Limit in days before the election up to which the coalitions percentages are calculated.
For example, specifying limitdays = 100 return coalitions percentages up to 100 days before the election.
If negative number is supplied (default value of -1), then data from all days is taken into account when calculating coalition percentages.}
}
\value{
dataframe of columns (date | Days to election | Coalition_1_percentage | ... | Coaltion_n_percentage)
  or visualisation points of the coalitions prediction.
}
\description{
Coalitions percentage prediction.
 
Provides prediction of percentages from different source types for specified coalitions. 
Coalitions percentages are predicted for each day in range [election_day - limit_days, election_day).
}
