% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{check_additional_args}
\alias{check_additional_args}
\title{checks additional arguments to add_data function}
\usage{
check_additional_args(newdata, pv, ...)
}
\arguments{
\item{newdata}{[\code{data.frame}]\cr
the data to add to the pollyvote object.}

\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the data to.}

\item{...}{additional named arguments to be checked.}
}
\value{
The pollyvote object with added data.
}
\description{
internal check function if additional arguments to pollyvoter::add_data are 
permissible for a pollyvote object
}
