% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{checks if data is permissible for a pollyvote object}
\usage{
check_data(newdata, pv)
}
\arguments{
\item{newdata}{[\code{data.frame}]\cr
the data to add to the pollyvote object.
should contain this columns: id, date, source, party, percent}

\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the data to.}
}
\value{
the checked data.
}
\description{
internal check function
}
\examples{
pv = create_pollyvote(perm_countries = "D")
data("election_result")
election_result = check_data(election_result, pv)

}
