% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalitions.R
\name{coalition_predictions_data}
\alias{coalition_predictions_data}
\title{Filters coalition predictions data and election_results in order to extract the election_date.}
\usage{
coalition_predictions_data(pv, election_year, limitdays)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr 
the pollyvote object.}

\item{election_year}{[\code{numeric(1)}]\cr
the election_year for which to get coalition predictions data.}

\item{limit_days}{[\code{numeric(1)}]\cr
limit in days before the election up to which the coalitions percentages are calculated.}
}
\value{
list of election_date and coalitions predictions data filtered based on the election_year and limit_days parameters.
}
\description{
Coalition predictions data from the pollyvote object is filtered based on 
the election_year and limit_days parameters.
}
