% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_names.R
\name{convert_names}
\alias{convert_names}
\title{convert names to lower character and ascii}
\usage{
convert_names(x)
}
\arguments{
\item{x}{[\code{character(n)}]\cr
character vector to be converted}
}
\value{
data.frame in long format
}
\description{
This is a function that converts a character vector to all lower case and ascii, see example.
}
\examples{
text = c("Grüne", "SPD")
convert_names(text)

}
