% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election_result}
\alias{election_result}
\title{"Election result" (BTW 2013)}
\format{A data frame with 8 rows and 2 variables:
\describe{
  \item{party}{Name of the party in all lower case.}
  \item{percent}{Share of the voters that are predicted to vote for this \code{party}}
}}
\usage{
election_result
}
\description{
A dataset in the long format containing election result for the BTW 2013 in Germany.
}
\keyword{datasets}

