% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_calc.R
\name{error_calc.pollyvote}
\alias{error_calc.pollyvote}
\title{calculate the prediction error of a pollyvote prediction.}
\usage{
\method{error_calc}{pollyvote}(object, method, ...)
}
\arguments{
\item{object}{[\code{pollyvote(1)}] \cr
pollyvote object to get the error calculation from.}

\item{method}{[\code{character(1)}] \cr
method name of the error calculation function of \code{object} to use.}

\item{...}{additional arguments to the error calculation function.}
}
\value{
the prediction
}
\description{
calculate the prediction error of a pollyvote prediction.
}
\seealso{
Other predict: \code{\link{error_calc}},
  \code{\link{predict.pollyvote}}
}
