% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{fill missing values in the data of a pollyvote object}
\usage{
fill_na(dat, na_handle = c("last", "omit", "mean_within", "mean_across"),
  pv = NULL, ...)
}
\arguments{
\item{dat}{[\code{data.frame()}]\cr
data frame of a pollyvote object.}

\item{na_handle}{[\code{character(1)}]\cr
character describing which method is used to fill missing values.
Currently supported is (TODO: give options and their meaning).}

\item{pv}{[\code{pollyvote}]\cr
only required for \code{na_handle = 'mean_across'}, the pollyvote object containing the data frame.}

\item{...}{[\code{list()}]\cr
additional arguments, currently ignored.}
}
\value{
the data set with filled mssing values.
}
\description{
internal function that handles missing values in the data frame of a pollyvote 
object. Different error handling arguments can be handed over.
}
