% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_colnames}
\alias{get_colnames}
\title{extract colnames}
\usage{
get_colnames(pv)
}
\arguments{
\item{pv}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
character containing colnames stored in \code{pv}.
}
\description{
This function extract colnames from a pollyvote container.
}
\examples{
pv = create_pollyvote()
# returns an empty data frame
get_colnames(pv)

}
