% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_countries}
\alias{get_countries}
\title{extract countries}
\usage{
get_countries(pv)
}
\arguments{
\item{pv}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
character vector containing all countries stored in \code{pv}.
}
\description{
This function extract countries from a pollyvote container.
}
\examples{
pv = create_pollyvote()
# returns an empty data frame
get_countries(pv)

}
