% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_date_latest}
\alias{get_date_latest}
\title{extract latest date}
\usage{
get_date_latest(pv)
}
\arguments{
\item{pv}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
character containing latest date stored in \code{pv}.
}
\description{
This function extract latest date from a pollyvote container.
}
\examples{
pv = create_pollyvote()
# returns an empty data frame
get_date_latest(pv)

}

