% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_election_date_from_election_year}
\alias{get_election_date_from_election_year}
\title{Get an election date that corresponds to a given \code{election_year}.}
\usage{
get_election_date_from_election_year(pv, election_year = NULL)
}
\arguments{
\item{pv}{[\code{pollyvote}] the pollyvote object.\cr}

\item{election_year}{[\code{numeric(1)}] the election year.
If it is [\code{NULL}] (default value), latest election_date is returned.}
}
\value{
An election date corresponding to the election_year.
}
\description{
Get an election date that corresponds to a given \code{election_year}.
}
