% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_election_result}
\alias{get_election_result}
\title{add an election result to a pollyvote object}
\usage{
get_election_result(pv, election, ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the election results to.}

\item{election}{[\code{character(1)}]\cr
name of the election. One of the permitted elections of \code{pv}.}

\item{...}{currently ignored.}
}
\value{
The pollyvote object with added prediction
}
\description{
Adds election result to a pollyvote opbject.
}
