% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_elections}
\alias{get_elections}
\title{extract elections}
\usage{
get_elections(id = "pollyvote")
}
\arguments{
\item{id}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
elections
}
\description{
This function extract elections from a pollyvote container.
}
\examples{
get_regions("pollyvote")

}
