% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_parties}
\alias{get_parties}
\title{extract party names}
\usage{
get_parties(id = "pollyvote")
}
\arguments{
\item{id}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
party names
}
\description{
This function extract party names from a pollyvote container.
}
\examples{
get_parties("pollyvote")

}
