% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_perm_colnames}
\alias{get_perm_colnames}
\title{extract colnames}
\usage{
get_perm_colnames(pv)
}
\arguments{
\item{pv}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
character containing colnames stored in \code{pv}.
}
\description{
This function extract colnames from a pollyvote container.
}
\examples{
pv = create_pollyvote(perm_countries = "D")
data("polls_individual")
pv = add_data(pv, newdata = polls_individual, country = "D", region = "national", 
              source_type = "poll", election = "BTW")
get_perm_colnames(pv)

}
