% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_perm_parties}
\alias{get_perm_parties}
\title{extract party names}
\usage{
get_perm_parties(pv)
}
\arguments{
\item{pv}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
character vector containing all party names stored in \code{pv}.
}
\description{
This function extract party names from a pollyvote container.
}
\examples{
pv = create_pollyvote()
# returns an empty data frame
get_perm_parties(pv)
 
}
