% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_perm_region_types}
\alias{get_perm_region_types}
\title{extract region types}
\usage{
get_perm_region_types(pv)
}
\arguments{
\item{pv}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
character vector containing all region types stored in \code{pv}.
}
\description{
This function extract region types from a pollyvote container.
}
\examples{
pv = create_pollyvote(perm_region_types = "national")
get_perm_region_types(pv)

}
