% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_perm_regions}
\alias{get_perm_regions}
\title{extract regions}
\usage{
get_perm_regions(pv)
}
\arguments{
\item{pv}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
character vector containing all regions stored in \code{pv}.
}
\description{
This function extract regions from a pollyvote container.
}
\examples{
pv = create_pollyvote(perm_regions = "national")
get_perm_regions(pv)

}
