% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_perm_source_types}
\alias{get_perm_source_types}
\title{extract source types}
\usage{
get_perm_source_types(pv)
}
\arguments{
\item{pv}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
character vector containing all source types stored in \code{pv}.
}
\description{
This function extract source types from a pollyvote container.
}
\examples{
pv = create_pollyvote(perm_source_types = "poll")
get_perm_source_types(pv)

}
