% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_perm_sources}
\alias{get_perm_sources}
\title{extract sources}
\usage{
get_perm_sources(pv)
}
\arguments{
\item{pv}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
character vector containing all sources stored in \code{pv}.
}
\description{
This function extract sources from a pollyvote container.
}
\examples{
pv = create_pollyvote(perm_sources = "poll")
get_perm_sources(pv)

}
