% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_region_weights}
\alias{get_region_weights}
\title{Get's the region weights from the pollyvote container.}
\usage{
get_region_weights(pv)
}
\arguments{
\item{pv}{[\code{pollyvote(1)}]\cr
the pollyvote object of which to extract the region weights from.}
}
\value{
the region region weights from the pollyvote container.
}
\description{
Get's the region weights from the pollyvote container.
}
\examples{
region_weights = get_region_weights(pv)

 
}
