% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_regions}
\alias{get_regions}
\title{extract regions}
\usage{
get_regions(pv)
}
\arguments{
\item{pv}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
character vector containing all regions stored in \code{pv}.
}
\description{
This function extract regions from a pollyvote container.
}
\examples{
pv = create_pollyvote()
# returns an empty data frame
get_regions(pv)

}

