% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_pollyvote_functions.R
\name{handle_region_method}
\alias{handle_region_method}
\title{Adds new column named 'electoral_result' in the input [\code{data}] parameter depending on the value of [\code{region_method}] parameter.}
\usage{
handle_region_method(data, region_method)
}
\arguments{
\item{data}{[\code{dataframe}]\cr
The predictions dataframe}

\item{region_method}{[\code{character(1)}]
Values:
 \itemize{
  \item [\code{character(1)}] wta: Winner takes it all method - 1 is assigned to the party which has most points in a region for a given day.
  \item [\code{character(1)}] vs:  Voteshares method - percent for the parties remain unchanged.
  }}
}
\value{
dataframe with 'electoral_result' column which a result of applying region aggregation method.
}
\description{
Adds new column named 'electoral_result' in the input [\code{data}] parameter depending on the value of [\code{region_method}] parameter.
}
