% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_exp_data_wide.R
\name{import_exp_data_wide}
\alias{import_exp_data_wide}
\title{import example data}
\usage{
import_exp_data_wide(sheet)
}
\arguments{
\item{sheet}{[\code{character(n)}]\cr
excel sheet name}
}
\value{
data.frame
}
\description{
This is a function that import example election data in 
wide format.
}
\examples{
elect_data_wide <- import_exp_data_wide("Wahlumfrage")


}

