% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_pollyvote_functions.R
\name{initial_prediction_aggr_source_type}
\alias{initial_prediction_aggr_source_type}
\title{initial aggregated source type prediction}
\usage{
initial_prediction_aggr_source_type(pv, which_source_type, agg_fun = "mean",
  na_handle = "last", ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object of which to get the prediction from.}

\item{which_source_type}{[\code{character(1)}]\cr
the name of the source to use for aggregation}

\item{agg_fun}{[\code{character(1)}]\cr
the name of the aggregation function to use, currently 'mean' and 'median' are supported}

\item{na_handle}{[\code{character(1)}]\cr
character describing which method is used to fill missing values.
Currently supported is 
\itemize{
\item \code{'last'}: Use the last available value. Fill all days between the earliest and latest day in the data.
\item \code{'omit'}: For every day, agregate available data.
\item \code{'mean_within'} Use the available data of this subcomponent to fill missing values.
\item \code{'mean_across'} Create a mini-pollyvote prediction by aggregating 
all \code{source_type}s using \code{na_handle = 'last'} and then
aggregating the result using \code{na_handle = 'last'}, which will then be used 
to predict missing values in the lowest aggregation level.
}}

\item{...}{[\code{list()}]\cr
additional arguments, currently ignored.}
}
\value{
data frame containing the results
}
\description{
TODO
}
