% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_pollyvote_functions.R
\name{initial_region_aggregation_pollyvote}
\alias{initial_region_aggregation_pollyvote}
\title{First calculates predictions on level of each region by using \code{initial_region_prediction_pollyvote} function
and afterwards aggregates predictions over the regions in order to obtain final single prediction for a party on given day.}
\usage{
initial_region_aggregation_pollyvote(pv, time_int = NULL, agg_fun = "mean",
  na_handle = "last", region_method = c("wta", "vs"))
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object of which to get the prediction data from.}

\item{agg_fun}{[\code{character(1)}]\cr
the name of the aggregation function to use, currently 'mean' and 'median' are supported}

\item{na_handle}{[\code{character(1)}]\cr 
specifies how NA values are handled in the data.}

\item{region_method}{method of aggregation of party scores over different regions.
See \code{handle_region_method} function for implementation and meaning of the parameteres.}

\item{time_int[\code{date}]}{the time interval for the prediction.}
}
\value{
dataframe containing aggregated predictions for the parties aggregated over regions.
}
\description{
First calculates predictions on level of each region by using \code{initial_region_prediction_pollyvote} function
and afterwards aggregates predictions over the regions in order to obtain final single prediction for a party on given day.
}
