% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_pollyvote_functions.R
\name{initial_region_prediction_pollyvote}
\alias{initial_region_prediction_pollyvote}
\title{Calculates prediction for each party on each day in each region from the data in the pollyvote container.}
\usage{
initial_region_prediction_pollyvote(pv, time_int = NULL, agg_fun = "mean",
  na_handle = "last", region = NULL)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object of which to get the prediction data from.}

\item{agg_fun}{[\code{character(1)}]\cr
the name of the aggregation function to use, currently 'mean' and 'median' are supported}

\item{na_handle}{[\code{character(1)}]\cr 
specifies how NA values are handled in the data.}

\item{region}{[\code{character(1)}] \cr
The region in which predictions are made. If the region is not existing in the pollyvote container, error is thrown.
If [\code{NULL}], then the predictions are calculated in each region.}

\item{time_int[\code{date}]}{the time interval for the prediction.}
}
\value{
dataframe of predictions for a party on each day in each region.
}
\description{
Calculates prediction for each party on each day in each region from the data in the pollyvote container.
}
