% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markets_eix}
\alias{markets_eix}
\title{"Eix" from source type "markets" (BTW 2013)}
\format{A data frame with 1104 rows and 3 variables:
\describe{
  \item{date}{Date of the poll.}
  \item{party}{Name of the party in all lower case.}
  \item{percent}{Share of the voters that are predicted to vote for this \code{party}}
}}
\usage{
markets_eix
}
\description{
A dataset in the long format containing election data ("Eix") from source type "markets" 
for the BTW 2013 in Germany.
}
\seealso{
\code{\link{markets_prognosys}}, \code{\link{markets_wahlfieber_1}}
}
\keyword{datasets}

