% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_pollyvote_functions.R
\name{moving_average_ci}
\alias{moving_average_ci}
\title{moving average}
\usage{
moving_average_ci(data, days_average = 7)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
data frame from ...}
}
\value{
data frame containing the results
}
\description{
Calculate moving average for confidence intervals.
}
