% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pollyvote.R
\name{plot.pollyvote}
\alias{plot.pollyvote}
\title{plot a pollyvote object}
\usage{
\method{plot}{pollyvote}(x, pred_name, ...)
}
\arguments{
\item{x}{[\code{pollyvote(1)}] \cr
pollyvote object to predict from.}

\item{pred_name}{[\code{character(1)}] \cr
name of the prediction to be plotted.}

\item{...}{currently unused.}
}
\value{
a ggplot object that can be further modified.
}
\description{
plots a prdiction of a pollyvote opbject.
}

