% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poll_data_long}
\alias{poll_data_long}
\title{individual polls of the BTW 2013}
\format{A data frame with 808 rows and 5 variables:
\describe{
  \item{id}{unique tag computed from \code{date} and \code{source}}
  \item{date}{Date of the poll.}
  \item{source}{Survey institute that conducted the survey.}
  \item{party}{Name of the party in all lower case.}
  \item{percent}{Share of the voters that are predicted to vote for this \code{party}}
}}
\usage{
poll_data_long
}
\description{
A dataset in the long format containing the individual polls of several survey institutes over 
the course of a few weeks for the BTW 2013 in Germany.
}
\seealso{
\code{\link{poll_data_wide}}
}
\keyword{datasets}

