% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poll_data_wide}
\alias{poll_data_wide}
\title{individual polls of the BTW 2013}
\format{A data frame with 101 rows and 11 variables:}
\usage{
poll_data_wide
}
\description{
A dataset in the wide format containing the individual polls of several survey institutes over 
the course of a few weeks for the BTW 2013 in Germany.
\describe{
  \item{id}{unique tag computed from \code{date} and \code{source}}
  \item{date}{Date of the poll.}
  \item{source}{Survey institute that conducted the survey.}
  \item{cdu/cdu, spd, grune, fdp, linke, piraten, afd, sonstige}{Share of votes for this party in percent}
}
}
\seealso{
\code{\link{poll_data_long}}
}
\keyword{datasets}

