% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polls_individual}
\alias{polls_individual}
\title{"Individual polls" from source type "polls" (BTW 2013)}
\format{A data frame with 808 rows and 5 variables:
\describe{
  \item{id}{unique tag computed from \code{date} and \code{source}}
  \item{date}{Date of the poll.}
  \item{source}{Survey institute that conducted the survey.}
  \item{party}{Name of the party in all lower case.}
  \item{percent}{Share of the voters that are predicted to vote for this \code{party}}
}}
\usage{
polls_individual
}
\description{
A dataset in the long format containing polls ("Individual polls") polls from source type "polls" 
for the BTW 2013 in Germany.
}
\seealso{
\code{\link{polls_individual}}

Other example.data: \code{\link{polls_wahlumfrage}}
}
\keyword{datasets}

