% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polls_wahlumfrage}
\alias{polls_wahlumfrage}
\title{"Wahlumfrage" from source type "polls" (BTW 2013)}
\format{A data frame with 1048 rows and 3 variables:
\describe{
  \item{date}{Date of the poll.}
  \item{party}{Name of the party in all lower case.}
  \item{percent}{Share of the voters that are predicted to vote for this \code{party}}
}}
\usage{
polls_wahlumfrage
}
\description{
A dataset in the long format containing polls ("Wahlumfrage") from source type "polls" 
for the BTW 2013 in Germany.
}
\seealso{
\code{\link{polls_individual}}, \code{\link{polls_tix}}

Other example data: \code{\link{polls_individual}}
}
\keyword{datasets}
