% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollyvoter.R
\docType{package}
\name{pollyvoter}
\alias{pollyvoter}
\alias{pollyvoter-package}
\title{pollyvoter: A package for aggregating polls and calculating forecasts}
\description{
The pollyvoter package provides methods for aggregating opinion polls,
calculating forecasts and determining error margins for elections results
}

