% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pollyvote.R
\name{predict.pollyvote}
\alias{predict.pollyvote}
\title{predict a pollyvote object}
\usage{
\method{predict}{pollyvote}(pv, name, ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the data to.}

\item{...}{additional arguments}
}
\value{
the prediction
}
\description{
predict a pollyvote opbject.
}

