% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalitions.R
\name{prediction_time_int}
\alias{prediction_time_int}
\title{Gets the prediction time interval based on the chosen election year.
The time interval should include the predictions data
after the previous election and target election.}
\usage{
prediction_time_int(pv, election_year)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr 
The pollyvote object}

\item{election_year}{[\code{numeric(1)}]\cr
The election year for which the coalitions are predicted.
If not specified, the most recent election year is used.}
}
\value{
prediction data time interval
}
\description{
Gets the prediction time interval based on the chosen election year.
The time interval should include the predictions data
after the previous election and target election.
}
