% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{set_data}
\alias{set_data}
\title{set data}
\usage{
set_data(pv, newdata)
}
\arguments{
\item{pv}{[\code{pollyvote(1)}]\cr
the pollyvote object of which to overwrite the data from.}
}
\value{
pollyvote object containing the new data.
}
\description{
This function completely overwrites the data from a pollyvote container.
Do not use unless you are sure that this is what you want.
}
\examples{
pv = create_pollyvote()
# returns an empty data frame
set_data(pv, newdata = data.frame(country = "D", election = "BTW"))

}
