% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{set_parties}
\alias{set_parties}
\title{set up party names}
\usage{
set_parties(parties, id = "pollyvote")
}
\arguments{
\item{parties}{[\code{character(1)}]\cr
party names}

\item{id}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
pollyvote container
}
\description{
This function set up party names in a pollyvote container.
}
\examples{
set_parties(parties = c("csu", "spd"), id = "pollyvote")

}
