% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{set_regions}
\alias{set_regions}
\title{set up regions}
\usage{
set_regions(regions, id = "pollyvote")
}
\arguments{
\item{regions}{[\code{character(1)}]\cr
regions}

\item{id}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
pollyvote container
}
\description{
This function set up regions in a pollyvote container.
}
\examples{
set_regions(regions = c("bayern", "bawu"), id = "pollyvote")

}

