% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{setParties}
\alias{setParties}
\title{Set parties to forecast}
\usage{
setParties(data = NULL, parties, remaining = NULL)
}
\arguments{
\item{data}{Data container. Creates a new data container if omitted}

\item{parties}{Vector of parties to forecast}

\item{remaining}{Character string naming the category for all unspecified
parties (e.g. "others"). If \code{NULL}, no category for other parties will
be created.}
}
\value{
Returns an updated data container
}
\description{
Set or change parties to forecast. This overwrites previously defined
parties.
}
\examples{
setParties(parties = c("Democrats", "Republicans"))
}
