% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalitions.R
\name{threshold_and_replace_party_with_coalition}
\alias{threshold_and_replace_party_with_coalition}
\title{Applies the threshold rule and replaces the parties in the election_data with coalitions names.}
\usage{
threshold_and_replace_party_with_coalition(data, threshold, threshold_handle,
  coalitions)
}
\arguments{
\item{data}{[\code{data.frame}]\cr 
the elections data frame.}

\item{threshold}{[\code{numeric(1)}]\cr
If positive, indicates the minimum voice component
that a party has to reach in order to participate in a coalition.}

\item{threshold_handle}{[\code{character(1)}]\cr
Specifies how to handle coalitions with parties that have less percentage than threshold.}

\item{coalitions}{[\code{list}]\cr
The coalitions for which prediction is calculated.}
}
\value{
data with applied threshold rule and replaced party names with coalition names.
}
\description{
For each day, it is checked whether data is available for all parties in the coalition.
If not, if threshold is greater than zero, NA or zero is inserted depending on the threshold_handle value.
}
