% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_to_long_format.R
\name{wide_to_long_format}
\alias{wide_to_long_format}
\title{convert election data set from wide to long format}
\usage{
wide_to_long_format(data)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
data frame in wide fromat containing TODO}
}
\value{
data.frame in long format
}
\description{
This is a function that converts an election data set from the
long  format into the wide format. Currently only German parties are supported.
}

