% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pollyvote.R
\name{write.pollyvote}
\alias{write.pollyvote}
\title{write a pollyvote prediction / error calculation}
\usage{
write.pollyvote(x, file = "", method = "write.table", prediction = NULL,
  error_calculation = NULL, write_args = list(), ...)
}
\arguments{
\item{x}{[\code{pollyvote(1)}] \cr
pollyvote object to predict from.}

\item{file}{[\code{character(1)}] \cr
either a character string naming a file or a connection open for writing. "" indicates output to the console.}

\item{method}{[\code{character(1)}] \cr
function name of the write function to use. Currently supported are
\code{'write.table'}, \code{'write.csv'} and \code{'write.csv2'}.}

\item{prediction}{[\code{character(1)}] \cr
method of the prediction to be plotted.}

\item{error_calculation}{[\code{character(1)}] \cr
method of the error_calculation to be plotted.}

\item{write_args}{[\code{list()}] \cr
list of arguments to the write function specified in \code{method}.}

\item{...}{additional arguments to the prediction or error calculation function.}
}
\value{
Nothing. A file is created.
}
\description{
writes the result of a prediction of a pollyvote object to a .csv file.
}
\examples{
pv = create_pollyvote(perm_countries = "D")
data("polls_individual")
pv = add_data(pv, newdata = polls_individual, country = "D", region = "national", 
              source_type = "poll", election = "BTW")
write.pollyvote(pv, file = "pred.csv", method = "write.table", 
                prediction = "pollyvote")

}
