% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_error_calc.R
\name{add_error_calc.pollyvote}
\alias{add_error_calc.pollyvote}
\title{add data to a pollyvote object}
\usage{
\method{add_error_calc}{pollyvote}(pv, name = "TODO", fun = function(pv) {  
    stop("TODO") }, ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the data to.}

\item{name}{[\code{character(1)}]\cr
name of the error calculation.}

\item{fun}{[\code{function(pv)}]\cr}

\item{...}{additional arguments}
}
\value{
The pollyvote object with added prediction
}
\description{
Adds data to a pollyvote opbject.
}

