% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prediction.R
\name{add_prediction.pollyvote}
\alias{add_prediction.pollyvote}
\title{add data to a pollyvote object}
\usage{
\method{add_prediction}{pollyvote}(pv, name = "no_aggregation",
  fun = function(pv) {     get_data(pv) }, ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the data to.}

\item{fun}{[\code{function(pv)}]\cr}

\item{...}{additional arguments}
}
\value{
The pollyvote object with added prediction
}
\description{
Adds data to a pollyvote opbject.
}

