% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prediction.R
\name{add_prediction}
\alias{add_prediction}
\title{add a prediciton (to a pollyvote object)}
\usage{
add_prediction(pv, name, fun = function(x) {     x }, ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the data to.}

\item{name}{[\code{character(1)}]\cr
name of the prediction.}

\item{fun}{[\code{function(pv)}]\cr}

\item{...}{additional arguments}
}
\value{
The pollyvote object with added prediction.
}
\description{
Adds a prediction, so far only implemented for pollyvote objects.
}

