% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pollyvote.R
\name{create_pollyvote}
\alias{create_pollyvote}
\title{create a pollyvote data container}
\usage{
create_pollyvote(id = "pollyvote", perm_countries = character(0),
  perm_sources = character(0), perm_source_types = character(0),
  perm_elections = character(0), perm_date_earliest = NULL,
  perm_date_latest = NULL, perm_regions = character(0),
  perm_region_types = character(0), perm_parties = character(0))
}
\arguments{
\item{id}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}

\item{perm_countries, perm_sources, perm_source_types, perm_elections, perm_regions, perm_region_types, perm_parties}{[\code{character(n)}]\cr
permitted values for the respective column of the data of this pollyvote object}

\item{perm_date_earliest, perm_date_latest}{[\code{POSIXct(1)}]\cr
earliest and latest permissible date in format \%Y-\%m-\%d.}
}
\value{
object of class pollyvote, containing an id, an initially empty data frame
  and potentially permissible values for the entries of the data frame.
}
\description{
A pollyvote container is a list containing all the information that is 
needed to predict an election. This is the central construct of this package 
around which all other functions are centered. 
A pollyvote container stores all relevant data and aggregation functions.
A pollyvote container can be 
predicted (\code{\link{predict.pollyvote}}) and its prediction error can be 
calculated (\code{\link{error_calc.pollyvote}}). The predictions and error 
calculations can also be plotted using \code{\link{plot.pollyvote}}.
}
\section{initial functions}{

  some useful prediction, aggregation and error calculation functions are already
  initialized when calling \code{create_pollyvote()}. 
  These are:
  \itemize{
    \item prediction functions
      \itemize{
      \item 'pollyvote': \code{\link{initial_prediction_pollyvote}}
      \item 'aggr_source_type'
    }
    \item error calculations: 'prediction_election'
    }
}

\examples{
pv = create_pollyvote(perm_countries = "D")

}
