% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_calc.R
\name{error_calc}
\alias{error_calc}
\title{calculate the prediction error of a pollyvote prediction.}
\usage{
error_calc(pv, name, ...)
}
\arguments{
\item{pv}{[\code{pollyvote(1)}] \cr
pollyvote object to predict from.}

\item{name}{[\code{character(1)}] \cr
name of the error calcluclation function on \code{pr} to use.}

\item{...}{additional arguments to the error calculation function.}
}
\value{
the prediction
}
\description{
calculate the prediction error of a pollyvote prediction.
}
\seealso{
Other predict: \code{\link{error_calc.pollyvote}},
  \code{\link{predict.pollyvote}}
}
