% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{fill missing values in the data of a pollyvote object}
\usage{
fill_na(dat, na_handle = c("last", "omit", "mean_within", "mean_across"),
  pv = NULL, time_int = NULL, ...)
}
\arguments{
\item{dat}{[\code{data.frame()}]\cr
data frame of a pollyvote object.}

\item{na_handle}{[\code{character(1)}]\cr
character describing which method is used to fill missing values.
Currently supported is 
\itemize{
\item \code{'last'}: Use the last available value. Fill all days between the earliest and latest day in the data.
\item \code{'omit'}: For every day, agregate available data.
\item \code{'mean_within'} Use the available data of this subcomponent to fill missing values.
\item \code{'mean_across'} Create a mini-pollyvote prediction by aggregating 
all \code{source_type}s using \code{na_handle = 'last'} and then
aggregating the result using \code{na_handle = 'last'}, which will then be used 
to predict missing values in the lowest aggregation level.
}}

\item{pv}{[\code{pollyvote}]\cr
only required for \code{na_handle = 'mean_across'}, the pollyvote object containing the data frame.}

\item{...}{[\code{list()}]\cr
additional arguments, currently ignored.}
}
\value{
the data set with filled mssing values.
}
\description{
internal function that handles missing values in the data frame of a pollyvote 
object. Different error handling arguments can be handed over.
}
