% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_data}
\alias{get_data}
\title{extract data}
\usage{
get_data(pv)
}
\arguments{
\item{pv}{[\code{pollyvote(1)}]\cr
the pollyvote object of which to extract the data from.}
}
\value{
data frame containing all the data stored in \code{pv}.
}
\description{
This function extracts the data from a pollyvote container.
}
\examples{
pv = create_pollyvote()
# returns an empty data frame
get_data(pv)

}

