% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_election_result}
\alias{get_election_result}
\title{get an election result from a pollyvote object}
\usage{
get_election_result(pv, election, ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object to add the election results to.}

\item{election}{[\code{character(1)}]\cr
name of the election. One of the permitted elections of \code{pv}.}

\item{...}{currently ignored.}
}
\value{
A data frame containing the election result.
}
\description{
gets an election result from a pollyvote object. To add a new election result to a pollyvote object
use \code{\link{add_election_result}}.
}
\examples{
pv = create_pollyvote(perm_countries = "D")
data("election_result")
pv = add_election_result(pv, "BTW 2013", election_result, date = "2013-09-22")
get_election_result(pv, election = "BTW 2013")

}
