% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_elections}
\alias{get_elections}
\title{extract elections}
\usage{
get_elections(pv)
}
\arguments{
\item{pv}{[\code{character(1)}]\cr
the name ID of the pollyvote object, defaults to 'pollyvote'.}
}
\value{
character vector containing all elections stored in \code{pv}.
}
\description{
This function extract elections from a pollyvote container.
}
\examples{
pv = create_pollyvote(perm_elections = "BTW")
# returns an empty data frame
get_elections(pv)

}

