% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_pollyvote_functions.R
\name{initial_error_calc_prediction_election}
\alias{initial_error_calc_prediction_election}
\title{initial pollyvote error calculation}
\usage{
initial_error_calc_prediction_election(pv, prediction = "pollyvote",
  target_election_year = NULL, ci = FALSE, alpha = 0.05, no_days = Inf,
  moving_average = TRUE, days_average = 7, ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object of which to get the prediction from.}

\item{prediction}{[\code{character(1)}]\cr
the name of the prediction function}

\item{target_election_year}{[\code{integer(1)}]\cr
the election year for which confidence intervals are calculated.
If \code{target_election_year} is \code{NULL}, CI are calculated for the last known election.}

\item{ci}{[\code{logical(1)}]\cr
whether confidence interval should be calculated}

\item{alpha}{[\code{numeric(1)}]\cr
significance level}

\item{no_days}{[\code{character(1)}]\cr}

\item{days_average}{length of moving average}

\item{...}{[\code{list()}]\cr
additional arguments, currently ignored.}
}
\value{
data frame containing the results
}
\description{
internal function which intializes a function for error calculation of method
\code{'prediction_election'}. This means, that one election result of the pollyvote
container is compared with the result of a prediction function of the same
pollyvote container.
}
