% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_pollyvote_functions.R
\name{initial_prediction_pollyvote}
\alias{initial_prediction_pollyvote}
\title{initial pollyvote prediction}
\usage{
initial_prediction_pollyvote(pv, agg_fun = "mean", na_handle = "last", ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object of which to get the prediction from.}

\item{agg_fun}{[\code{character(1)}]\cr
the name of the aggregation function to use, currently 'mean' and 'median' are supported}

\item{na_handle}{[\code{character(1)}]\cr
character describing which method is used to fill missing values.
Currently supported is (TODO: give options and their meaning).}

\item{...}{[\code{list()}]\cr
additional arguments, currently ignored.}
}
\value{
data frame containing the results
}
\description{
TODO
}
