% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_pollyvote_functions.R
\name{initial_prediction_pollyvote}
\alias{initial_prediction_pollyvote}
\title{initial pollyvote prediction}
\usage{
initial_prediction_pollyvote(pv, time_int = NULL, agg_fun = "mean",
  na_handle = "last", ...)
}
\arguments{
\item{pv}{[\code{pollyvote}]\cr
the pollyvote object of which to get the prediction data from.}

\item{agg_fun}{[\code{character(1)}]\cr
the name of the aggregation function to use, currently 'mean' and 'median' are supported}

\item{na_handle}{[\code{character(1)}]\cr 
specifies how NA values are handled in the data.}

\item{...}{[\code{list()}]\cr
additional arguments, currently ignored.}

\item{time_int[\code{date}]}{the time interval for which the prediction.}
}
\value{
data frame containing predictions for a parties in the days before the election
}
\description{
internal function which initializes the prediction function of method pollyvote.
This function aggregates the daat in two steps: 
In the first step all predictions from sources of the same \code{source_types} are aggregated
daily. In the second step the aggregated \code{source_types} are aggregated,
resulting in one prediction per day and party.
}
