% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pollyvote.R
\name{plot.pollyvote}
\alias{plot.pollyvote}
\title{plot a pollyvote object}
\usage{
\method{plot}{pollyvote}(x, .prediction_method = NULL,
  .error_calc_method = NULL, ...)
}
\arguments{
\item{x}{[\code{pollyvote(1)}] \cr
pollyvote object to predict from.}

\item{.prediction_method}{[\code{character(1)}] \cr
method of the prediction to be plotted.}

\item{.error_calc_method}{[\code{character(1)}] \cr
method of the error calculation to be plotted. Note that only one of prediction_method 
and error_calc_method can be specified.}

\item{...}{additional arguments to the predict or error_calc_method function}
}
\value{
a ggplot object that can be further modified.
}
\description{
plots a prediction of a pollyvote object.
}
