% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pollyvote.R
\name{plot.pollyvote}
\alias{plot.pollyvote}
\title{plot a pollyvote object}
\usage{
\method{plot}{pollyvote}(x, prediction = NULL, error_calc = NULL, ...)
}
\arguments{
\item{x}{[\code{pollyvote(1)}] \cr
pollyvote object to predict from.}

\item{prediction}{[\code{character(1)}] \cr
name of the prediction to be plotted.}

\item{error_calc}{[\code{character(1)}] \cr
name of the error calculation to be plotted. Note that only one of prediction 
and error_calc can be specified.}

\item{...}{additional arguments to the predict or error_calc function}
}
\value{
a ggplot object that can be further modified.
}
\description{
plots a prdiction of a pollyvote opbject.
}
