% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pollyvote.R
\name{predict.pollyvote}
\alias{predict.pollyvote}
\title{predict a pollyvote object}
\usage{
\method{predict}{pollyvote}(object, method, ...)
}
\arguments{
\item{object}{[\code{pollyvote(1)}] \cr
pollyvote object to predict from.}

\item{method}{[\code{character(1)}]\cr
method of the prediction. This method name will be used to call the prediction in \code{\link{predict.pollyvote}}.}

\item{...}{additional arguments}
}
\value{
a data frame containing the result of the prediction
}
\description{
predict a pollyvote opbject.
}
\examples{
pv = create_pollyvote(perm_countries = "D") 
data("polls_individual")
pv = add_data(pv, newdata = polls_individual, country = "D", region = "national", 
              source_type = "poll", election = "BTW")
pred <- predict(pv, method = "pollyvote")               
              

}
\seealso{
Other predict: \code{\link{error_calc.pollyvote}},
  \code{\link{error_calc}}
}
