% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pollyvote.R
\name{predict.pollyvote}
\alias{predict.pollyvote}
\title{predict a pollyvote object}
\usage{
\method{predict}{pollyvote}(object, name, ...)
}
\arguments{
\item{object}{[\code{pollyvote(1)}] \cr
pollyvote object to predict from.}

\item{name}{[\code{character(1)}]\cr
name of the prediction.}

\item{...}{additional arguments}
}
\value{
the prediction
}
\description{
predict a pollyvote opbject.
}
\seealso{
Other predict: \code{\link{error_calc.pollyvote}},
  \code{\link{error_calc}}
}

