% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pollyvote.R
\name{print.pollyvote}
\alias{print.pollyvote}
\title{print a pollyvote data container}
\usage{
\method{print}{pollyvote}(pv)
}
\arguments{
\item{pv}{a pollyvote object.}
}
\description{
internal method to print out the most important features of a pollyvote container
}
