% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalitions.R
\name{valid_coalitions}
\alias{valid_coalitions}
\title{Checks whether coalitions are made of permitted parties.
Permitted parties can be defined either in pv$permparties or in allowed_parties parameter.
Therefore, the idea is first to collect all specified permitted parties in one vector.
This vector will serve as a source for checking whether parties inside the coalitions have valid names.
If there are no parties specified, then the check for valid party names is not performed.}
\usage{
valid_coalitions(coalitions, pv, permitted_parties)
}
\arguments{
\item{pv}{the pollyvote object.}

\item{input}{coalitions to the function.}

\item{allowed_parties}{the input allowed_parties object.}
}
\value{
list of coalitions consisted only of permitted parties.
}
\description{
Checks whether coalitions are made of permitted parties.
Permitted parties can be defined either in pv$permparties or in allowed_parties parameter.
Therefore, the idea is first to collect all specified permitted parties in one vector.
This vector will serve as a source for checking whether parties inside the coalitions have valid names.
If there are no parties specified, then the check for valid party names is not performed.
}
